Fix tests and docs
