+++ tests/service/test_import_service.py
@@ -1,24 +1,19 @@
-import types
 import unittest
 from unittest import mock
 
 from exabel_data_sdk import ExabelClient
 from exabel_data_sdk.client.api.data_classes.entity import Entity
-from exabel_data_sdk.scripts.create_entities_from_csv import CreateEntitiesFromCsv
+from exabel_data_sdk.service.import_service import CsvImportService
 
 
 class TestCreateEntitiesFromCsv(unittest.TestCase):
     def test_create_entities_with_description(self):
-        args = [
-            "script-name",
-            "--filename-input",
-            "./tests/resources/data/entities_with_description.csv",
-        ]
 
-        script = CreateEntitiesFromCsv(args, "Create entities")
         client = mock.create_autospec(ExabelClient(host="host", api_key="123"))
         client.entity_api.entity_exists.return_value = False
-        result = script.run_script(client, script.parse_arguments())
+        service = CsvImportService(client)
+        result = service.create_entities_from_csv(
+            filename_input="./tests/resources/data/entities_with_description.csv", separator=";")
 
         # first row
         # entityTypes/brand/test.brand_y;brand_y;brand_y description
@@ -46,6 +41,26 @@ class TestCreateEntitiesFromCsv(unittest.TestCase):
             result["entityTypes/brand/entities/test.BRAND_X"], None, "Result not as expected"
         )
 
+    def test_create_entities_with_existing_entity(self):
+
+        client = mock.create_autospec(ExabelClient(host="host", api_key="123"))
+        client.entity_api.entity_exists.return_value = True
+        service = CsvImportService(client)
+        result = service.create_entities_from_csv(
+            filename_input="./tests/resources/data/entities_with_description.csv", separator=";")
+
+        # first row
+        # entityTypes/brand/test.brand_y;brand_y;brand_y description
+        # fails with ValueError
+        self.assertEqual(result["entityTypes/brand/test.brand_y"], None, "Result not as expected")
+
+        # second row
+        # entityTypes/brand/entities/test.BRAND_X;BRAND_X;BRAND_X description
+        self.assertEqual(client.entity_api.create_entity.call_args_list, [], "Unexpected call")
+        self.assertTrue(
+            "entityTypes/brand/entities/test.BRAND_X" not in result, "Result not as expected"
+        )
+
     def test_create_entities_without_description(self):
         args = [
             "script-name",
@@ -53,12 +68,11 @@ class TestCreateEntitiesFromCsv(unittest.TestCase):
             "./tests/resources/data/entities_without_description.csv",
         ]
 
-        script = CreateEntitiesFromCsv(args, "Create entities")
         client = mock.create_autospec(ExabelClient(host="host", api_key="123"))
         client.entity_api.entity_exists.return_value = False
-        result = script.run_script(client, script.parse_arguments())
-
-        print(f"result = {result}")
+        service = CsvImportService(client)
+        result = service.create_entities_from_csv(
+            filename_input="./tests/resources/data/entities_without_description.csv", separator=";")
 
         # first row
         # entityTypes/brand/entities/test.BRAND_X;BRAND_X
